Move validator row and charts to validators directory
