--- app/javascript/packs/stake_accounts/charts/chart_variables.js
@@ -1,10 +1,0 @@
-export default {
-  chart_darkgrey: '#979797',
-  chart_lightgrey: '#cacaca',
-  chart_green: '#00ce99',
-  chart_blue: '#0091f2',
-  chart_green_t: 'rgba(0, 206, 153, 0.4)',
-  chart_blue_t: 'rgba(0, 145, 242, 0.25)',
-  chart_lightgrey_t: 'rgba(202, 202, 202, 0.3)',
-  chart_font_size: 11
-}
